EventManager/requirements.txt
flask
flask-bootstrap
flask-login
flask-mail
flask-sqlalchemy
sqlalchemy-migrate
flask-wtf
pymysql
